Task 1 [NUMBER GAME]
The game generates a random number between 1 and 100 at the start of each round. If your guess is incorrect, you can continue guessing until you find the right number, with feedback provided after each guess. To add a layer of challenge, there is a limit on the number of attempts you have to guess the number correctly. Your score is calculated based on the number of attempts taken. The fewer attempts you make, the higher your score will be. At the end of the game, your total score will be displayed.